Predictions ui update & user history section

+++ backend/controllers/predictions.controller.js
@@ -62,7 +62,11 @@ export const uploadAndPredict = async (req, res) => {
 
     if (newPrediction) {
       const prediction = await newPrediction.save();
-      await client.del(`userHistory:${userId}`);
+      const history = await client.get(`userHistory:${userId}`);
+      if (history) {
+        await client.del(`userHistory:${userId}`);
+      }
+
       res.status(201).json(prediction);
     } else {
       res.status(400).json({ error: "Invalid parameters" });
@@ -101,12 +105,10 @@ export const getMyHistory = async (req, res) => {
 
     const cachedHistory = await client.get(`userHistory:${userId}`);
     if (cachedHistory) {
-      console.log(cachedHistory);
       return res.status(200).json(JSON.parse(cachedHistory));
     }
 
     const predictions = await Prediction.find({ userId: userId });
-    console.log(predictions.length);
     await client.set(`userHistory:${userId}`, JSON.stringify(predictions));
     await client.expire(`userHistory:${userId}`, 30 * 24 * 60 * 60);
     res.status(200).json(predictions);
